requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langgraph>=0.0.20

# Search API
tavily-python>=0.3.0

# PDF Generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: Development Dependencies
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
